Update README.md
I watched a ted talk the other day. This Ted Talk was about the falsely conceived notion that something great comes from a single slice of genius. Linda Hill explained that significant innovation is born from many collective minds. To get this point across Linda used an experience she had from studying Pixar Animation. To make the movie Ratatouille, 250 people were involved in the story writing and animation. Every single person added a slice of their genius in order to make the movie whole. An example of this is one of the animators who decided to add an arched eyebrow in order to exhibit malicious intent. This subtle difference, though slightly, impacted the movie in a positive way. This is what Ms. Hill means by “You have to unleash the talents and you have to harness them into a work that is actually useful”. While I agree with Linda’s point, I also think that there are special cases. For example, Jeff Bezos started his company, Amazon, in a garage by himself, but to Linda’s argument, it is true that many more great ideas are done through the work of many.